requirements.txt
pandas==2.3.3
python-dotenv
scikit-learn
seaborn